teams-bot/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# HTTP Client
httpx==0.25.2
aiofiles==24.1.0

# Teams Bot Framework
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0
aiohttp==3.9.1

# Configuration & Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==23.2.1

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.8

# Azure Integration
azure-identity==1.15.0